Student Enrollment System
This project is designed to develop an online platform for students to
enroll courses for a particular semester. The technologies used are Java
programming language, Spring MVC framework, Oracle 11g database. The
project requirement is based on identifying the current problems in the
existing process. As a result, the designed web application serves as
student friendly and provides the flexibility to the students as well as
the faculty in managing the enrollment system.